Refactor favorite login on a component to use in movie page
